+++ examples/nmeadump.py
@@ -2,14 +2,17 @@
 Simple command line utility to stream the parsed NMEA output of an NMEA GNSS device.
 
 Usage (all args are optional):
-nmeadump.py port="COM6" baud=9600 timeout=5 nmea_only=0 raw=0
+nmeadump.py port="COM6" baud=9600 timeout=5 nmea_only=0 validate=1 raw=0
 
 If nmea_only=True (1), streaming will terminate on any non-NMEA data.
+If validate & 1, will check for valid checksum (otherwise will ignore during reading, 
+but possibly fail later during parsing if message is corrupt).
+If validate & 2, will check for valid msgId (otherwise will ignore).
 """
 
 import sys
 from serial import Serial
-from pynmeagps.nmeareader import NMEAReader
+from pynmeagps import NMEAReader, GET, VALCKSUM
 
 PORT = "COM6"
 BAUD = 9600
@@ -26,13 +29,14 @@ def stream_ubx(**kwargs):
         baud = int(kwargs.get("baud", BAUD))
         timeout = int(kwargs.get("timeout", TIMEOUT))
         nmea_only = int(kwargs.get("nmea_only", 0))
+        validate = int(kwargs.get("validate", VALCKSUM))
         rawformat = int(kwargs.get("raw", 0))
         print(
             f"\nStreaming from {port} at {baud} baud in",
             f"{'raw' if rawformat else 'parsed'} format...\n",
         )
         stream = Serial(port, baud, timeout=timeout)
-        nmr = NMEAReader(stream, nmea_only)
+        nmr = NMEAReader(stream, nmea_only, validate, GET)
         for (raw, parsed) in nmr:
             if rawformat:
                 print(raw)
@@ -51,7 +55,7 @@ if __name__ == "__main__":
                 "the parsed NMEA output of an NMEA GNSS device.\n\n",
                 "Usage (all args are optional): nmeadump.py",
                 'port="COM6" baud=9600 timeout=5',
-                "nmea_only=0 raw=0\n\n Type Ctrl-C to terminate.",
+                "nmea_only=0 validate=1, raw=0\n\n Type Ctrl-C to terminate.",
             )
             sys.exit()