+++ examples/nmeafile.py
@@ -7,7 +7,7 @@ Created on 7 Mar 2021
 @author: semuadmin
 """
 
-from pynmeagps.nmeareader import NMEAReader
+from pynmeagps import NMEAReader, VALCKSUM, GET
 import pynmeagps.exceptions as ube
 
 
@@ -57,14 +57,14 @@ class NMEAStreamer:
                 print(f"Error closing file {err}")
         self._connected = False
 
-    def reader(self, validate=False, mode=0):
+    def reader(self, nmea_only=False, validate=VALCKSUM, mode=GET):
         """
         Reads and parses NMEA message data from stream
         using NMEAReader iterator method
         """
 
         i = 0
-        self._ubxreader = NMEAReader(self._stream, validate, mode)
+        self._ubxreader = NMEAReader(self._stream, nmea_only, validate, mode)
 
         for msg in self._ubxreader:  # invokes iterator method
             try:
@@ -85,19 +85,22 @@ if __name__ == "__main__":
 
     print("Enter fully qualified name of file containing binary NMEA data: ", end="")
     filefqn = input().strip('"')
-    print("Do you want to validate the data stream (y/n)? (n) ", end="")
-    val = input() or "n"
-    VALD = val in ("Y", "y", "YES,", "yes", "True")
+    print("Do you want to ignore non-NMEA data (y/n)? (y) ", end="")
+    val = input() or "y"
+    NMEA_ONLY = val in ("N", "n", "NO,", "no", "False")
+    print("Do you want to validate the data stream (0/1/2/3)? (1) ", end="")
+    val = input() or "1"
+    VALD = int(val)
     print("Message mode (0=GET (output), 1=SET (input), 2=POLL (poll)? (0) ", end="")
-    mode = input() or "0"
-    MODED = int(mode)
+    moded = input() or "0"
+    MODED = int(moded)
 
     print("Instantiating NMEAStreamer class...")
     ubf = NMEAStreamer(filefqn)
     print(f"Opening file {filefqn}...")
     ubf.open()
     print("Starting file reader")
-    ubf.reader(VALD, MODED)
+    ubf.reader(NMEA_ONLY, VALD, MODED)
     print("\n\nClosing file...")
     ubf.close()
     print("Test Complete")