+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # pynmeagps Release Notes
 
+### RELEASE 0.1.3-beta
+
+1. Parse validate flag added to NMEAReader.read() method, defaults to VLCKSUM (1 - check checksum only); other options are VALNONE = 0, 
+VALMSGID = 2. Options can be ANDed. Examples updated accordingly.
+
 ### RELEASE 0.1.2-beta
 
 1. Constructor enhanced to allow NMEAMessage to be created using typed values.